Created MapFactory and Region classes, moved City class to maps package, removed map creation from America class, implemented Color class in Region to remove REGION_* constants from BaseMap
